Split login components out into a few more files for readability
It was getting hard to work with that huge login.js file
