+++ app/renderer/components/login.js
@@ -1,10 +1,13 @@
 import electron from 'electron';
 import React from 'react';
-import Blockies from 'react-blockies';
-import {If, autoBind} from 'react-extras';
+import {autoBind} from 'react-extras';
 import Api from '../api';
+import CreatePortfolioButton from './create-portfolio-button';
+import PortfolioItem from './portfolio-item';
 
 /* eslint-disable */
+
+const {getPortfolios, decryptSeedPhrase} = electron.remote.require('./portfolio-util');
 
 const initMarketmaker = seedPhrase => new Promise(resolve => {
 	electron.ipcRenderer.send('start-marketmaker', {seedPhrase});
@@ -30,191 +33,6 @@ const initApi = async seedPhrase => {
 	});
 };
 
-const portfolio = electron.remote.require('./portfolio');
-
-const PortfolioImage = ({onClick, ...rest}) => (
-	<div className="PortfolioImage" onClick={onClick}>
-		<Blockies
-			{...rest}
-			size={10}
-			scale={6}
-			bgColor="transparent"
-			color="rgba(255,255,255,0.15)"
-			spotColor="rgba(255,255,255,0.25)"
-		/>
-	</div>
-);
-
-class CreatePortfolio extends React.Component {
-	constructor(props) {
-		super(props);
-		autoBind(this);
-
-		this.state = {showPortfolioForm: false};
-	}
-
-	showPortfolioForm() {
-		this.setState({showPortfolioForm: true});
-	}
-
-	hidePortfolioForm() {
-		this.setState({showPortfolioForm: false});
-	}
-
-	async onSubmit(event) {
-		event.preventDefault();
-		this.hidePortfolioForm();
-		await portfolio.create(this.state);
-		this.props.loadPortfolios();
-	}
-
-	render() {
-		return (
-			<div>
-				<button className="add-portfolio btn btn-sm btn-primary btn-block" onClick={this.showPortfolioForm} disabled={this.state.showPortfolioForm}>Add portfolio</button>
-				<If condition={this.state.showPortfolioForm} render={() => (
-					<div className="add-portfolio-modal modal-dialog">
-						<div className="modal-content">
-							<form className="portfolio-form" onSubmit={this.onSubmit}>
-								<div className="modal-header">
-									<h5 className="modal-title">Add portfolio</h5>
-									<button type="button" className="close" onClick={this.hidePortfolioForm}>
-										<span>&times;</span>
-									</button>
-								</div>
-								<div className="modal-body">
-									<div className="form-group">
-										<input
-											type="text"
-											className="form-control"
-											placeholder="Portfolio Name"
-											onChange={e => this.setState({ name: e.target.value })}
-											autoFocus
-										/>
-									</div>
-									<div className="form-group">
-										<input
-											type="text"
-											className="form-control"
-											placeholder="Seed Phrase"
-											onChange={e => this.setState({ seedPhrase: e.target.value })}
-										/>
-									</div>
-									<div className="form-group">
-										<input
-											type="password"
-											className="form-control"
-											placeholder="Password"
-											onChange={e => this.setState({ password: e.target.value })}
-										/>
-									</div>
-								</div>
-								<div className="modal-footer">
-									<button type="submit" className="btn btn-primary">Add</button>
-									<button type="button" className="btn btn-secondary" onClick={this.hidePortfolioForm}>Close</button>
-								</div>
-							</form>
-						</div>
-					</div>
-				)}/>
-			</div>
-		);
-	}
-}
-
-class Portfolio extends React.Component {
-	constructor(props) {
-		super(props);
-		autoBind(this);
-
-		this.state = {
-			isLoginInputVisible: this.props.showLoginForm,
-			isCheckingPassword: false,
-		};
-	}
-
-	showLoginInput() {
-		this.setState({
-			isLoginInputVisible: true,
-		});
-	}
-
-	async onSubmit(event) {
-		event.preventDefault();
-
-		this.setState({isCheckingPassword: true});
-
-		const {encryptedSeedPhrase} = this.props.portfolio;
-		const password = this.input.value;
-
-		try {
-			// TODO: Show some loading here as it takes some time to decrypt the password and then start marketmaker
-			const seedPhrase = await portfolio.decryptSeedPhrase(encryptedSeedPhrase, password);
-			const api = await initApi(seedPhrase);
-
-			// TODO: These should be defaults saved in the config and changeable by the user
-			await Promise.all([
-				api.enableCoin('KMD'),
-				api.enableCoin('VTC'),
-				api.enableCoin('LTC'),
-			]);
-
-			const {portfolio: currencies} = await api.portfolio();
-
-			this.props.setAppState({
-				portfolio: this.props.portfolio,
-				currencies,
-				api
-			});
-		} catch (err) {
-			console.error(err);
-
-			this.input.value = '';
-
-			const passwordError = /Authentication failed/.test(err.message) ? "Incorrect password" : err.message;
-			this.setState({
-				isCheckingPassword: false,
-				passwordError,
-			});
-		}
-	}
-
-	render() {
-		const {portfolio} = this.props;
-
-		return (
-			<div className="Portfolio">
-				<PortfolioImage seed={portfolio.fileName} bgColor="transparent" onClick={this.showLoginInput} />
-				<h4>{portfolio.name}</h4>
-				<If condition={this.state.isLoginInputVisible} render={() => (
-					<form className="login-form" onSubmit={this.onSubmit}>
-						<div className="form-group">
-							<input
-								type="password"
-								className="form-control"
-								placeholder="Enter Your Password"
-								ref={input => this.input = input}
-								disabled={this.isCheckingPassword}
-								autoFocus
-							/>
-						</div>
-						<div className="form-group" disabled={this.isCheckingPassword}>
-							<button type="submit" className="btn btn-primary btn-sm btn-block">Login</button>
-						</div>
-						<If condition={Boolean(this.state.passwordError)} render={() => (
-							<div className="form-group">
-								<div className="alert alert-danger" role="alert">
-									{this.state.passwordError}
-								</div>
-							</div>
-						)}/>
-					</form>
-				)}/>
-			</div>
-		);
-	}
-}
-
 export default class Login extends React.Component {
 	constructor(props) {
 		super(props);
@@ -229,16 +47,38 @@ export default class Login extends React.Component {
 
 	async loadPortfolios() {
 		this.setState({
-			portfolios: await portfolio.getAll(),
+			portfolios: await getPortfolios(),
+		});
+	}
+
+	async handleLogin(portfolio, password) {
+		// TODO: Show some loading here as it takes some time to decrypt the password and then start marketmaker
+		const seedPhrase = await decryptSeedPhrase(portfolio.encryptedSeedPhrase, password);
+		const api = await initApi(seedPhrase);
+
+		// TODO: These should be defaults saved in the config and changeable by the user
+		await Promise.all([
+			api.enableCoin('KMD'),
+			api.enableCoin('VTC'),
+			api.enableCoin('LTC'),
+		]);
+
+		const {portfolio: currencies} = await api.portfolio();
+
+		this.props.setAppState({
+			portfolio,
+			currencies,
+			api
 		});
 	}
 
 	render() {
 		const portfolios = this.state.portfolios.map(portfolio => (
-			<Portfolio
+			<PortfolioItem
 				key={portfolio.fileName}
 				portfolio={portfolio}
 				showLoginForm={this.state.portfolios.length === 1}
+				handleLogin={this.handleLogin}
 				{...this.props}
 			/>
 		));
@@ -254,7 +94,7 @@ export default class Login extends React.Component {
 
 		return (
 			<div className="Login container">
-				<CreatePortfolio loadPortfolios={this.loadPortfolios} />
+				<CreatePortfolioButton loadPortfolios={this.loadPortfolios} />
 				{portfolioContainer}
 			</div>
 		);
